Implemented lastRan increments when adding/deleting members
Addresses #1, closes #2
Altered the delFusionMember function to output an array of successfully removed members rather than just the number of successfully deleted members. This enables passing to getNewLastRanValue_ in order to obtain the lastRan change
The corner case of deleting a member when they are also the lastRan reference remains unaddressed, with the default behavior of not altering the lastRan parameter.
Updated getNewLastRanValue to handle calls when the next database update is a scoreboard write.