FT bugfix: handle add/del if lastRan >= numMembers
If the scoreboard is ready to be written, lastRan will have a value greater than the number of members. If a query for the member located at lastRan is made, an empty result is returned, which throws a TypeError when trying to set the original UID of the member located at lastRan pre-add / pre-del.
Added a sanity check that only cares about shifting lastRan if we are in the middle of an update cycle.